requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
defusedxml==0.7.1
dj-rest-kit==1.1.1
Django==4.2.11
django-allauth==0.57.0
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.5
django-jazzmin==2.6.0
django-json-widget==1.1.1
django-lifecycle==1.1.2
django-model-utils==4.3.1
django-role-permissions==3.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-extra-fields==3.7.0
drf-writable-nested==0.7.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
filetype==1.2.0
future==1.0.0
geographiclib==2.0
geopy==2.4.1
idna==3.7
inflection==0.5.1
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
phonenumbers==8.13.29
pillow==10.2.0
postfinancecheckout==5.1.0
psycopg==3.1.18
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
